remove a test from TCK
This test requires sorting by fields which are not selected
but as for as I know, the specification does not explicitly
require this, and it is something that is not usually
considered allowed, though it's true that many databases are
quite forgiving here. Now, it's true that I didn't disallow
this in JDQL, but there's a big difference between allowing
it statically and allowing it for dynamically-added sorting.